Make PojoFromJsonPayloadUnmarshaller work with `Type`
It previously used `Class` which meant it couldn't be used with parameterized types such as `List<T>`. In a previous commit, we added support for `Type` in `JsonLibrary`. This continues this refactoring. Since it's no longer restricted to work with POJOs, it has been renamed to `JsonPayloadUnmarshaller`.

`ListPayloadUnmarshaller` was removed as it serves no purpose anymore. Tests for `List` payloads have been added to JsonPayloadUnmarshallerTest.

`PayloadUnmarshaller#getTargetClass` was refactored to return `Type` as well. This fixes user output when using primitive types such as byte arrays. The output for the type was `[B` before, now it's `byte[]`. This is easier to understand for users without deeper JVM knowledge.