Make PojoAsJsonFunctionResultMarshaller work with `Type`
Similar to a previous commit where unmarshalling was changed to use `Type`. `PojoAsJsonFunctionResultMarshaller` was renamed `JsonFunctionResultMarshaller` since it now works with more than just POJOs.

This adds terrible exception handling in `SalesforceFunctionsProjectFunctionsScanner` for determining the output type for `List` types, copied from the unmarshalling section. Both will be fixed in a later commit.